docs(repo.cheat): add commands to navigate to project path and project from anywhere in repo
First command allows you to navigate to the project path
given a project name by grepping the manifest file.

Second command allows you to navigate to a project from
anywhere in the repo using the fzf tool.
It lists all the project paths and allows you to select
one using a preview of the files in the project.